Split out the profile sharing components
