+++ src/components/app/MenuButtons/index.css
@@ -36,128 +36,6 @@
   background-color: rgba(0, 0, 0, 0.1);
 }
 
-.menuButtonsCompositeButtonContainer {
-  display: flex;
-  flex-flow: column nowrap;
-  border-right: 1px solid var(--grey-30);
-  border-left: 1px solid var(--grey-30);
-}
-
-.menuButtonsShareButton,
-.menuButtonsUploadingButton,
-.menuButtonsPermalinkButton,
-.menuButtonsUploadErrorButton,
-.menuButtonsSecondaryShareButton {
-  height: 24px;
-  margin-bottom: -24px;
-}
-
-.menuButtonsUploadingButton {
-  overflow: hidden;
-}
-
-.menuButtonsUploadingButtonInner {
-  display: flex;
-  overflow: hidden;
-  height: 100%;
-  flex-flow: column nowrap;
-  align-items: stretch;
-}
-
-.menuButtonsUploadingButtonLabel {
-  position: relative;
-  padding: 0 10px;
-  color: hsla(0, 0%, 100%, 0.7);
-  cursor: default;
-  line-height: 24px;
-  text-align: center;
-  -moz-user-select: none;
-}
-
-/* The button width needs to be large enough so that the text "Share Without URLs" fits */
-.currentButtonIsPermalinkButton.currentButtonIsSecondaryShareButton
-  .menuButtonsPermalinkButton,
-.menuButtonsUploadingButtonInner,
-.menuButtonsPermalinkButtonButton,
-.menuButtonsUploadErrorButtonButton,
-.menuButtonsSecondaryShareButtonButton {
-  width: 115px;
-}
-
-/* The button will be larger during symbolication to fit the text:
-"Sharing will be enabled once symbolication is complete" */
-.menuButtonsShareButtonButton {
-  min-width: 115px;
-}
-
-.menuButtonsUploadingButtonProgress {
-  width: 100%;
-  height: 24px;
-  margin-bottom: -24px;
-  border: 0;
-  background: var(--internal-uploading-button-background-color);
-  transform-origin: bottom left;
-}
-
-.menuButtonsUploadingButtonProgress::-moz-progress-bar {
-  background: var(--internal-uploading-progress-fill-color);
-}
-
-.menuButtonsShareButtonButton,
-.menuButtonsSecondaryShareButtonButton {
-  background-color: #0d8730;
-  color: white;
-}
-
-/* CSSTransition */
-.buttonWithPanelButton,
-.menuButtonsUploadingButtonInner {
-  transition: opacity 200ms ease-in-out, transform 200ms ease-in-out;
-}
-
-.menuButtonsTransitionUp-enter .buttonWithPanelButton,
-.menuButtonsTransitionUp-enter .menuButtonsUploadingButtonInner {
-  opacity: 0;
-  transform: translateY(100%);
-}
-
-.menuButtonsTransitionUp-enter-active .buttonWithPanelButton,
-.menuButtonsTransitionUp-enter-active .menuButtonsUploadingButtonInner {
-  opacity: 1;
-  transform: translateY(0);
-}
-
-.menuButtonsTransitionUp-exit .buttonWithPanelButton,
-.menuButtonsTransitionUp-exit .menuButtonsUploadingButtonInner {
-  opacity: 1;
-  transform: translateY(0);
-}
-
-.menuButtonsTransitionUp-exit-active .buttonWithPanelButton,
-.menuButtonsTransitionUp-exit-active .menuButtonsUploadingButtonInner {
-  opacity: 0;
-  transform: translateY(-100%);
-}
-
-/* end of CSSTransition */
-
-/* permalink button is hidden whenever uploading is in progress, but is
- * moved up when uploading is finished. */
-.currentButtonIsUploadingButton .menuButtonsPermalinkButtonButton {
-  opacity: 0;
-  transform: translateY(100%);
-}
-
-.currentButtonIsPermalinkButton.currentButtonIsSecondaryShareButton
-  .menuButtonsSecondaryShareButton {
-  margin-left: 115px;
-}
-
-.menuButtonsPrivacyPanel,
-.menuButtonsUploadErrorPanel {
-  --width: 280px;
-}
-
 .menuButtonsProfileDownloadPanel {
   --offset-from-right: 25px;
   --width: 30em;
@@ -165,23 +43,4 @@
 
 .menuButtonsDownloadLink {
   white-space: nowrap;
-}
-
-.menuButtonsPermalinkTextField {
-  width: 14em;
-  height: 19px;
-  padding: 0 4px;
-  margin: 5px -5px;
-}
-
-.menuButtonsShareNetworkUrlsContainer {
-  margin: 20px 0;
-  -moz-user-select: none;
-  user-select: none;
-}
-
-.menuButtonsShareNetworkUrlsCheckbox {
-  position: relative;
-  bottom: 1px;
-  vertical-align: middle;
 }
